change: enable hc is optional
