+++ pkg/common/server/middlewares.go
@@ -44,26 +44,3 @@ func (s Server) setUserID() gin.HandlerFunc {
 		c.Next()
 	}
 }
-
-// healthz is a liveness probe
-func (s *Server) healthz() gin.HandlerFunc {
-	return func(ctx *gin.Context) {
-		ctx.Status(http.StatusNoContent)
-	}
-}
-
-// readyz is a readiness probe
-func (s *Server) readyz() gin.HandlerFunc {
-	return func(ctx *gin.Context) {
-		if !s.readyzAvailable.Load().(bool) {
-			http.Error(ctx.Writer, http.StatusText(http.StatusServiceUnavailable), http.StatusServiceUnavailable)
-			return
-		}
-		ctx.Status(http.StatusNoContent)
-	}
-}
-
-// SetReadyzState enables/disables readyz
-func (s *Server) SetReadyzState(state bool) {
-	s.readyzAvailable.Store(state)
-}
